vote: rename vote file
